FigureFactory 구조체 구현
