+++ CoordinateCalculator/CoordinateCalculator/Figure.swift
@@ -8,6 +8,11 @@
 
 import Foundation
 
+enum FigureType: Int {
+    case point = 1
+    case line = 2
+}
+
 protocol Figure {
     
 }
