+++ CoordinateCalculator/CoordinateCalculator/InputError.swift
@@ -11,4 +11,5 @@ import Foundation
 enum InputError: String, Error {
     case NonInputStandard = "소괄호로 둘러싸인 x,y 좌표를 정확히 입력해주세요. ex: (4,16)"
     case NonCoordinateStandard = "x,y 좌표는 0부터 24까지 입력해주세요."
+    case UndefinedFigure = "현재 지원하지 않는 도형입니다."
 }
